support support for analysis of huge files
In VS Code, when opening files larger than 50MB, the use of extensions is
restricted to ensure performance and memory efficiency. This limitation
helps maintain a responsive and stable environment when handling large
files. You can bypass this limitation and use log-analysis on large files
by the vsc-lfs extension. This tool enables the handling of large files
in VS Code, allowing you to apply log-analysis capabilities.

Signed-off-by: KimHyoJoong <46334592+JeanTracker@users.noreply.github.com>